FIXED distributed input splitting
Client-streaming for input-splitting, now fully functional.

Removed synchronised method, as each of the 4 latches are independent of each others and their counting is not printed / used elsewhere. Concurrency although exists, it does not impact the output blocks / outcome.

Client-streaming was working previously; However, the real issue was that the ArrayList.add(position, ..) method only inserted and shifted the position, instead of replacing. This led to indexOutOfBounds exceptions when accessing each tile within that ArrayList. Replaced with, ArrayList.set(position, ..), which is currently in use.
It is likely some previous versions of the client-streaming communication pattern could have been actually working, if the set(..) method were to be used and the ArrayList outputBlocksArray was initialised / it was added 4 positions before calling set(..)